Changes for AudioIntelligenceService

+++ DriveAudioRecorder/src/com/agreeya/audiodetectionapp/AudioChunk.java
@@ -1,11 +1,11 @@
 package com.agreeya.audiodetectionapp;
 
 public class AudioChunk {
-	byte[] data;
+	short[] data;
 	int length = 0;
 	int used = 0;
 	
 	public AudioChunk(int size) {
-		data = new byte[size];
+		data = new short[size];
 	}
 }
